Normalization with develop (src, some test)

+++ test/swmr_sparse_reader.c
@@ -29,7 +29,6 @@
 /* Headers */
 /***********/
 
-#include <inttypes.h> /* for PRIu64 */
 #include "h5test.h"
 #include "swmr_common.h"
 
@@ -118,7 +117,7 @@ check_dataset(hid_t fid, unsigned verbose, const symbol_info_t *symbol, symbol_t
                   (uintmax_t)start[0], (uintmax_t)start[1]);
 
     /* Read record from dataset */
-    record->rec_id = UINT64_MAX;
+    record->rec_id = (uint64_t)ULLONG_MAX;
     if (H5Dread(dsid, symbol_tid, rec_sid, file_sid, H5P_DEFAULT, record) < 0)
         return -1;
